chore(configs): rename eslint to lint
