+++ packages/configs/src/configs/lint.js
